separated approved hours list into distinct page
