+++ DoGoodBackend/.gitignore
@@ -35,3 +35,7 @@ out/
 
 ### VS Code ###
 .vscode/
+
+
+## Evil Firebase File
+!**/src/main/resources/serviceAccountKey.json
